oldVersions/requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyserial==3.5
pywin32-ctypes==0.2.3
setuptools==80.9.0
